Add Web Playback SDK to React app to connect to Spotify as a Web Player (#12)
The SDK Web Playback was added to the react file to connect to Spotify however somethings were left out. Validation of the token key from Spotify needs to be run through the props as well as hidden from user information. The 'ts-ignores' are temporary to allow the code to run. Issues will be created for both separately. 

closes issue #5