Displaying Information From Spotify Player to User (#14)
* clicking functionality with button

the button will be used to display information about current player object when clicked.

* State Hook Problem Review

Implementing the useState hook to store and update the player object. Method not permanent.

* Review Commit

Current_track.artists[0].name is added to line 50 in WebPlayBack.tsx two issues arise.
1: Multiple instances of Player Object.
2: Deconstructing TypeScript Error.

* finished functionality to issue #6

An interface was created so that individualized states could be consolidated. isPlayerGood was removed with playerinfo for less state usage. Button was there for testing and has now been removed for build.

* Comments

This is a commit after a pull request but before resolution.

* Pull Request Changes

See Pull Request 14 for info on why these changes are needed.