Player Controls (#15)
* clicking functionality with button

the button will be used to display information about current player object when clicked.

* State Hook Problem Review

Implementing the useState hook to store and update the player object. Method not permanent.

* Review Commit

Current_track.artists[0].name is added to line 50 in WebPlayBack.tsx two issues arise.
1: Multiple instances of Player Object.
2: Deconstructing TypeScript Error.

* finished functionality to issue #6

An interface was created so that individualized states could be consolidated. isPlayerGood was removed with playerinfo for less state usage. Button was there for testing and has now been removed for build.

* Player Controls

This is an implemented controller system for beginning player controls. The team created a Backward, Toggle, and forward button for easy to start with player functionality.

* Fixed PR issues

Removed access token for saftey, and capsulated buttons into separate PlayerUI component.