Favicon and Title Update (#29)
We decided on a music note from the heart. This project has always been more about the heart. Represented in how we really try to learn from a fundamental level what we are doing and how to replicate it.

We decided on 'MusicFree' as a 'Working Title' for now. A themed website would be rather corny, but would be the most fun to build. Like Moosic. Where everything is Cow themed. (for example).

This PR resolves Issue: #22