Styling the User Controls (#32)
* Styling the User Controls

The styling of the User Controls was a lot more difficult than we anticipated. The middle slider is to be replaced by the song position slider which will need styling. The middle slider is the volume control and will need to be moved when the song position is done.

* update for PR Comments

Tech Lead Miles had left a comment stating I had a comment that was unnecessarily commenting.