Organized the search results
We displayed the search results in square like sections of album, artist, and track. After styling the image we styled the text. Finally we created a button that when clicked sent information to the Spotify API requesting that the player play the album, artist, or track from the specified device_id. This device Id we sent through props and perhaps should have sent through Re-dux.