Displayed current Track Album and Artist on left of screen (#35)
* Displayed current Track Album and Artist on left of screen

We decided to get the outline of these UI components before getting in the greater detail. This PR is an example of that strategy.

* Changed updated for PR

We decided having Uniform Quotations across the platform would be best.

* PR request changes

The AppBar component is rewriting the imported css code so the AppBar is inline for now.