Organized Search Result (#33)
* Organized Search Result

We organized the search results into Material UI elements leaving room for CardActions. Like a button pressed that plays the song associated to card, or album or artist. This is in line 28 of file 'SearchResultItem.tsx'.

This Resolves issue #24

* Organized the search results

We displayed the search results in square like sections of album, artist, and track. After styling the image we styled the text. Finally we created a button that when clicked sent information to the Spotify API requesting that the player play the album, artist, or track from the specified device_id. This device Id we sent through props and perhaps should have sent through Re-dux.

* PR Commit for requested changes

There is a few comment changes.

Co-authored-by: MtuMartian <mrmartin@mtu.edu>